Add better handling of the IDL session exiting
If the IDL session exits, this detects the EOF exception thrown by
pexpect and converts it into a regular IOError exception that can be
caught by the calling program, and also prints an informative message
that the IDL session has exited.

This allows better pIDLy control for the case of an IDL session that
consists of e.g. running some pre-existing procedure that does some
particular calculation and then exits.